Use doubles in actSelPress calculation

+++ hw05-0036505665/src/main/java/hr/fer/zemris/optjava/dz5/part1/GeneticAlgorithm.java
@@ -143,7 +143,7 @@ public class GeneticAlgorithm {
                 newPopulation.add(randomSelection.from(pool));
             }
 
-            actSelPress = (newPopulation.size() + pool.size()) / population.size();
+            actSelPress = (double) (newPopulation.size() + pool.size()) / population.size();
             population = newPopulation;
         }
